+++ src/Index.js
@@ -2,11 +2,12 @@ import React from 'react'
 import { Table } from './table/Table'
 import * as R from 'ramda'
 import CreateButton from './CreateButton'
+import { FilterModal, FilterModalButton, isTableFilterable } from './table/Filter'
 import {
   getActions,
   getHasIndex,
-  getIndexFields,
-  getModelLabelPlural
+  getIndexFields, getModelLabel,
+  getModelLabelPlural, getSingleton
 } from './utils/schemaGetters'
 import { Redirect } from 'react-router-dom'
 import {
@@ -17,19 +18,42 @@ import {
   skipOverride
 } from './Utils'
 
-export const DefaultIndexTitle = ({ schema, modelName, path, data, user, customProps }) => {
+export const DefaultIndexTitle = ({
+  schema,
+  modelName,
+  selectOptions,
+  path,
+  data,
+  user,
+  tableView,
+  customProps
+}) => {
   const actions = getActions(schema, modelName)
   const onCreateClick = R.path(['create', 'onIndexCreate'], actions)
   const onClick = () => onCreateClick({ modelName, path })
   const creatable = isCreatable({ schema, modelName, data, user, customProps })
+  const filterable = isTableFilterable({ schema, modelName })
+  const currentFilters = R.path(['filter', modelName], tableView)
+  const filterOrder = R.path(['filterOrder', modelName], tableView)
+  const filtersAreActive = R.path(['filtersAreActive', modelName], tableView)
+
   return (
     <div style={{ marginBottom: '10px' }}>
       <h3 className='d-inline'>
         {getModelLabelPlural({schema, modelName, data, user, customProps })}
       </h3>
-      {creatable && <div className='float-right'>
-        <CreateButton {...{ onClick }} />
-      </div>}
+      {filterable && <FilterModal {...{
+        schema,
+        modelName,
+        selectOptions,
+        data,
+        filterOrder,
+        filterInputs: currentFilters
+      }}/>}
+      <div className='float-right'>
+        {filterable && <FilterModalButton {...{ modelName, filtersAreActive }} />}
+        {creatable && <CreateButton {...{ onClick }} />}
+      </div>
     </div>
   )
 }
@@ -41,10 +65,11 @@ const DefaultIndex = ({
   modalData,
   editData,
   selectOptions,
+  modelStore,
   path,
   tooltipData,
   user,
-  tableOptions,
+  tableView,
   customProps
 }) => {
   if (!getHasIndex(schema, modelName)) {
@@ -86,7 +111,7 @@ const DefaultIndex = ({
             path,
             tooltipData,
             user,
-            tableOptions,
+            tableView,
             customProps
           }}
         />
@@ -100,10 +125,11 @@ const DefaultIndex = ({
             modalData,
             editData,
             selectOptions,
+            modelStore,
             path,
             tooltipData,
             user,
-            tableOptions,
+            tableView,
             customProps,
             fieldOrder,
             fromIndex: true,
@@ -123,16 +149,39 @@ const Index = ({
   modalData,
   editData,
   selectOptions,
+  modelStore,
   path,
   tooltipData,
   user,
-  tableOptions,
+  tableView,
   customProps
 }) => {
+  // if singleton, Index redirects to Detail pg
+  if (getSingleton(schema, modelName)) {
+    const singleton = R.last(data)
+    // singleton may not be null when last deleted; test for 'id'
+    const singleId = R.propOr(null, 'id', singleton)
+    if (singleId) {
+      return <Redirect to={`/${modelName}/${singleId}`} />
+    }
+    // if no singleId exists, must create
+    const actions = getActions(schema, modelName)
+    const onCreateClick = R.path(['create', 'onIndexCreate'], actions)
+    return (
+      <div className='container'>
+        <h1>
+          {`No ${getModelLabel({ schema, modelName, data, user, customProps })} Exists`}
+          <CreateButton {...{
+            onClick: () => onCreateClick({ modelName })
+          }} />
+        </h1>
+      </div>
+    )
+  }
+
   const IndexOverride = getIndexOverride(schema, modelName)
-
   const IndexComponent = IndexOverride || DefaultIndex
-
+  
   return skipOverride(IndexOverride) ? null : (
     <IndexComponent
       {...{
@@ -142,10 +191,11 @@ const Index = ({
         modalData,
         editData,
         selectOptions,
+        modelStore,
         path,
         tooltipData,
         user,
-        tableOptions,
+        tableView,
         customProps
       }}
     />